# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Atlas API",
        "description": "Documentation interactive de l'API Atlas",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/v1/availabilities/{agentId}": {
            "get": {
                "tags": [
                    "Disponibilités"
                ],
                "summary": "Consulter les disponibilités d'un agent",
                "operationId": "af053bc4ee895b6b1f30302d521ff247",
                "parameters": [
                    {
                        "name": "agentId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des disponibilités"
                    },
                    "404": {
                        "description": "Agent non trouvé"
                    }
                }
            }
        },
        "/api/v1/availabilities": {
            "post": {
                "tags": [
                    "Disponibilités"
                ],
                "summary": "Ajouter une disponibilité",
                "operationId": "6b6e8d39c6e72e549fd3ca87123a66ee",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "start_date",
                                    "end_date"
                                ],
                                "properties": {
                                    "start_date": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-05-01T10:00:00Z"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-05-01T12:00:00Z"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Disponibilité créée"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "422": {
                        "description": "Chevauchement avec un créneau existant"
                    }
                }
            }
        },
        "/api/v1/availabilities/{id}": {
            "delete": {
                "tags": [
                    "Disponibilités"
                ],
                "summary": "Supprimer une disponibilité",
                "operationId": "961702a56c0dae9202acb26c5229676f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Disponibilité supprimée"
                    },
                    "403": {
                        "description": "Action non autorisée"
                    },
                    "404": {
                        "description": "Disponibilité non trouvée"
                    }
                }
            }
        },
        "/api/v1/bookings": {
            "get": {
                "tags": [
                    "Réservations"
                ],
                "summary": "Liste des réservations de l'utilisateur",
                "operationId": "2f7e7260c708227e4c44de9d5b59ea36",
                "responses": {
                    "200": {
                        "description": "Liste des réservations"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                }
            },
            "post": {
                "tags": [
                    "Réservations"
                ],
                "summary": "Demander une réservation",
                "operationId": "9ec8aa3de0d1efdd0cfe72c27eac745c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "agent_profile_id",
                                    "slot_start",
                                    "slot_end"
                                ],
                                "properties": {
                                    "agent_profile_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "slot_start": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-05-01T10:00:00Z"
                                    },
                                    "slot_end": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-05-01T12:00:00Z"
                                    },
                                    "message": {
                                        "type": "string",
                                        "example": "Je souhaite visiter le centre historique"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Réservation créée"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "422": {
                        "description": "L'agent n'est pas disponible"
                    }
                }
            }
        },
        "/api/v1/bookings/{id}/accept": {
            "post": {
                "tags": [
                    "Réservations"
                ],
                "summary": "Accepter une réservation",
                "operationId": "2b787f856dae74bb7dda53fa693dd3ec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Réservation acceptée"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "404": {
                        "description": "Réservation non trouvée"
                    }
                }
            }
        },
        "/api/v1/bookings/{id}/refuse": {
            "post": {
                "tags": [
                    "Réservations"
                ],
                "summary": "Refuser une réservation",
                "operationId": "526b92bdcba6139b6fcfd9043abc01ed",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Réservation refusée"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "404": {
                        "description": "Réservation non trouvée"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}